Move XOs to plugin/ui
